add keydow and up eventlistner on document and created keyboardHandler component in util
