+++ src/store/keyboard/context.js
@@ -1,44 +1,18 @@
-import { createContext, useContext, useState, useEffect, useRef } from 'react';
-import { ALT_KEY } from '@constants/AppConstants';
+import { createContext, useContext, useState } from 'react';
 
 const KeyboardContext = createContext();
 
-function isOptionKey(e) {
-  return e.key === ALT_KEY;
-}
-
 export const KeyboardProvider = ({ children }) => {
   const [isOptionKeyPressed, setIsOptionKeyPressed] = useState(false);
-  const divRef = useRef(null);
 
   const initialValue = {
     isOptionKeyPressed,
+    setIsOptionKeyPressed,
   };
-
-  useEffect(() => {
-    if (!divRef) return;
-    divRef.current.focus();
-  }, []);
 
   return (
     <KeyboardContext.Provider value={initialValue}>
-      <div
-        tabIndex={-1}
-        role="presentation"
-        ref={divRef}
-        onKeyDown={(e) => {
-          if (isOptionKey(e)) {
-            setIsOptionKeyPressed(true);
-          }
-        }}
-        onKeyUp={(e) => {
-          if (isOptionKey(e)) {
-            setIsOptionKeyPressed(false);
-          }
-        }}
-      >
-        {children}
-      </div>
+      {children}
     </KeyboardContext.Provider>
   );
 };
